"Validation Added"

+++ JavaWork/SpringBootApi/src/main/java/com/springBootApi/SpringBootApi/service/imple/StudentServiceImple.java
@@ -1,6 +1,7 @@
 package com.springBootApi.SpringBootApi.service.imple;
 
 import com.springBootApi.SpringBootApi.entity.Student;
+import com.springBootApi.SpringBootApi.exception.StudentNotFoundException;
 import com.springBootApi.SpringBootApi.repository.StudentRepository;
 import com.springBootApi.SpringBootApi.service.StudentService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
 public class StudentServiceImple implements StudentService {
@@ -28,18 +30,31 @@ public class StudentServiceImple implements StudentService {
     }
 
     @Override
-    public Student fetchStudentById(long id) {
-        return studentRepository.findById(id).get();
+    public Student fetchStudentById(long id) throws StudentNotFoundException {
+        Optional<Student> student=studentRepository.findById(id);
+        if(student.isEmpty()){
+            throw new StudentNotFoundException(id);
+        }
+        return student.get();
     }
 
     @Override
-    public void deleteStudentById(long id) {
+    public void deleteStudentById(long id) throws StudentNotFoundException {
+        Optional<Student> student=studentRepository.findById(id);
+        if(student.isEmpty()){
+            throw new StudentNotFoundException(id);
+        }
         studentRepository.deleteById(id);
     }
 
     @Override
-    public Student updateStudent(long id, Student student) {
-        Student student1=studentRepository.findById(id).get();
+    public Student updateStudent(long id, Student student) throws StudentNotFoundException {
+        Optional<Student> stud=studentRepository.findById(id);
+
+        if(stud.isEmpty()){
+            throw new StudentNotFoundException(id);
+        }
+        Student student1=stud.get();
         if(Objects.nonNull(student.getFirstName()) && !"".equalsIgnoreCase(student.getFirstName())){
             student1.setFirstName(student.getFirstName());
         }
